logic updte
logic update
